+++ app/controllers/dashboards_controller.rb
@@ -2,9 +2,11 @@
 
 # Controller for dashboard
 class DashboardsController < ApplicationController
-  before_action :set_user_events, only: [:index]
+  before_action :set_user_events, only: [:events]
 
   def index; end
+
+  def events; end
 
   private