+++ app/controllers/events_controller.rb
@@ -28,7 +28,7 @@ class EventsController < ApplicationController
 
   def destroy
     @event.destroy
-    redirect_to root_path
+    redirect_to events_path
   end
 
   private