+++ spec/system/dashboards_spec.rb
@@ -13,12 +13,13 @@ RSpec.describe 'Dashboards', type: :system do
 
     it 'lists user events' do
       visit dashboard_path
+      page.find(:id, 'page').click_link 'My Events'
       row_count = get_event_rows(page).size
       expect(row_count).to eq(5)
     end
 
     it 'deletes event on click' do
-      visit dashboard_path
+      visit events_path
       rows = get_event_rows(page)
       rows.first.click_on('Delete Event')
       row_count = get_event_rows(page).size